requirements.txt
fastapi==0.70.0
uvicorn==0.15.0
tensorflow-macos==2.12.0
Pillow==8.2.0
